Re-structure react app Add Persons component Add CockPit component
