--- src/App.css
@@ -1,33 +1,0 @@
-.App {
-  text-align: center;
-}
-
-.App input{
-  margin-top: 20px;
-}
-.red{
-  color:red;
-}
-.bold{
-  font-weight: bold;
-}
-.App button{
-  border: 1px solid blue;
-  background-color: green;
-  font: inherit;
-  color: white;
-  padding: 10px;
-  cursor: pointer;
-}
-
-.App button:hover{
-  background-color: lightblue;
-  color: black;
-}
-.App button.Red{
-  background-color: red;
-}
-.App button.Red:hover{
-  background-color: salmon;
-  color: black;
-}
