+++ src/components/CharComp/CharComp.css
