+++ src/components/CharComp/CharComp.js
