+++ src/containers/App.css
@@ -1,0 +1,4 @@
+.App {
+  text-align: center;
+}
+
