+++ src/containers/App.js
@@ -1,0 +1,74 @@
+import React, { Component } from 'react';
+
+import classes from  './App.css';
+import Persons from '../components/Persons/Persons';
+import CockPit from '../components/Cockpit/Cockpit';
+
+
+class App extends Component {
+  state = {
+    persons: [
+      { id: 'asfa1', name: 'Max', age: 28 },
+      { id: 'vasdf1', name: 'Manu', age: 29 },
+      { id: 'asdf11', name: 'Stephanie', age: 26 }
+    ],
+    otherState: 'some other value',
+    showPersons: false
+  }
+
+  nameChangedHandler = ( event, id ) => {
+    const personIndex = this.state.persons.findIndex( p => {
+      return p.id === id;
+    } );
+
+    const person = {
+      ...this.state.persons[personIndex]
+    };
+
+    person.name = event.target.value;
+
+    const persons = [...this.state.persons];
+    persons[personIndex] = person;
+
+    this.setState( { persons: persons } );
+  }
+
+  deletePersonHandler = ( personIndex ) => {
+    const persons = [...this.state.persons];
+    persons.splice( personIndex, 1 );
+    this.setState( { persons: persons } );
+  }
+
+  togglePersonsHandler = () => {
+    const doesShow = this.state.showPersons;
+    this.setState( { showPersons: !doesShow } );
+  }
+
+  render () {
+  
+    let persons = null;
+    if ( this.state.showPersons ) {
+      persons = (
+         <Persons
+         clicked={this.deletePersonHandler}
+         changed={this.nameChangedHandler}
+         persons={this.state.persons}
+         />
+      );
+    }
+
+
+    return (
+      <div className={classes.App}>
+        <CockPit 
+        persons = {this.state.persons}
+        showPersons = {this.state.showPersons}
+        toggle = {this.togglePersonsHandler}
+        />
+        {persons}
+     </div>
+    );
+  }
+}
+
+export default App;
